Refactor: BudgetRepository 를 수정한다
RepositoryTest를 상속받아서 구현
